requirements.txt
flask==3.0.0
praw==7.7.1
python-dotenv==1.0.0
anthropic==0.39.0
requests==2.31.0
gunicorn==21.2.0
